Update patch set 2
Patch Set 2:

Btw, I think I mentioned this but any of your code (I realize it's not all the code) needing to run in the certain bootloaders is going to have a bad time with this logging infrastructure ... this is because UEFI's printf() is non-standard e.g. %s means a UCS-2 string and %a is for ASCII. This is why libavb only has avb_debug(const char* str) and avb_debugv(const char *str, ...).

Patch-set: 2
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>