ese-replay: transcieve replay tool
This tool provide a very simple session replay
function. It reads in a file of the format:
  hex... [hex..]\n
  ...
The first hexadecimal string is the transmission to
send. The second is an optional expected reply which will
be checked against the end of the response.  That
allows for an easy check for 9000 on a successful APDU.

More work is needed to clean up this code.  In practice, this
tool and ese-relay should be unified and rolled into the
C++ interface.

Test: live device testing
Bug: 35102956

Change-Id: Ic598fcf2dc4e7ebf28cf3ad75446f8b63d31874e