libese-hw: move transceive to scattergather
This change migrates hardware callbacks for transceive to a
scatter-gather interface, updates the tests, and adds specific tests for
ese_sg.

Additionally, a small amount of code reorganization is included, like
moving API to ESE_API and into sysdeps to reduce some duplicatation as
well as removing dead code (two sysdeps.h).

As the T=1 implementation already copies data during framing, it was
becoming ungainly to recopy data to frame APDUs which will then be
copied only to be reframed.  This is especially true for extended APDUs
as the payloads may be up to 65k.

Change-Id: I37f586645965d32577f8b724dbe41c9041439887
Test: ese_replay still works (open channel, select, close channel)
Bug: 34467857