[draft] apps/boot: boot applet, library, and tool.
This is a placeholder.  The applet provides a sample interface, but
it isn't actually functional (compiles, installs, selects, but not much
more).  The tool and library setup a logical channel and communicate
with the applet and some initial unittests are in place with a stubbed
C++ Ese hardware instance.

The boot library interfaces are still a bit clunky. A pass through
the functionality to unify storage and interfacing needs to happen.

Test: this doesn't work yet.
Bug: 34460238,34467857,34684505

Change-Id: I64223548b5c95f24044eb98c9f999468c73bcf4a