apps/boot: boot applet, library, and tool.
This change covers a lot of ground (sorry).

It provides an applet and three packages for use on an NXP smartcard
with some additional APIs.  The applet and packages provide secure
storage for bootloader use:
- rollback indices
- cascading locks (carrier->owner)

Rollback index storage is implemented in VersionStorage.java.

Applet state (pre-production/production) and hardware state
(inBootloader) are provided via GlobalStateImpl.java.  This file is an
implementation of the globalstate.OwnerInterface, a shareable interface
which gets its own .cap file (via OwnerInterface.java).  The client
interface is implemented in CallbackInterface.java and is primarily used
to allow the primary applet to indicate a data clearing operation to
other applets.

Storage.java provides the overall Applet interface.

All the locks are implemented in CarrierLock.java and BasicLock.java.
Each lock can only be toggled if the preceding lock in the locks array
is unlocked.

The parent lock, CarrierLock, relies on a (test, in this CL) key for
RSA_SHA256 PKCS#1 padded signature over a device specific hash and a
server provided rollback nonce (monotonically increasing 64-bit value).

With the applets, is an ant build.xml.  As building applets is not
trivially supported in the Android build system, they can be built by
running 'ant' in the same directory as the build file.  This will fetch
any necessary third party dependencies, then build the cap files.  The
globalstate*.caps must be installed before avb_storage.cap.

This change also includes an interface library built on top of libese
and an example command line tool for using the applet.  The tool and
library setup a logical channel and communicate with the applet and some
initial unittests are in place with a stubbed C++ Ese hardware instance.

The library errors codes and applet status codes need to be refined

Test: all functions tested manually with the tool and ese-replay except
      for carrier test vectors/data collection/verification. That is
      still outstanding work.
Bug: 34460238,34467857,34684505

Change-Id: I64223548b5c95f24044eb98c9f999468c73bcf4a