apps/boot: Support streaming APDU payloads.
Locks with larger metadata payloads can now stream them
to a scratch buffer using a new command set.  This allows
the library and applet to easily be adjusted to different
hardware capabilities.

This change updates both the library and the applet as
well as the applet version.

(This change increases the flash footprint by 2k, but
attempts to avoid actually writing to it.)

Test: submit a 4k key and it writes and reads.
Bug: 38150381
Change-Id: I0759db7388f8f42a7828b699b7b05b6046ec5a53