Support uint16/uint32 for ENUM Tag.
This commit introduces support for using uint16/uint32 as the underlying
data type for an enum tag. Prior to this change, the enum tag was
limited to uint8 data type. With this enhancement, uint16 and uint32 can
be used, allowing for a wider range of values to be used as enum tags.

Bug: 291356939
Test: run vts -m VtsAidlKeyMintTarget
Change-Id: I6a6fe9b7023fb8868e0a0f311e3075256532fdb5