Revert "Disable -Wno-single-bit-bitfield-constant-conversion locally"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/927f1f80-4c82-4688-8efb-a16a39cc7d90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362820028
Change-Id: I3dbc4f90004ba75d82f47b8cc378966649bcf136
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com